requirements.txt
faster-whisper>=1.0.0
soundcard>=0.4.0
silero-vad>=5.0
numpy>=1.24.0
pyyaml>=6.0
opencc-python-reimplemented>=0.1.7
